bit-tutorial/ui/login/login.spec.tsx
import React from 'react';
import { render, fireEvent, screen } from '@testing-library/react';
import { BasicLogin } from './login.composition';

describe('Login', () => {
  it('renders correctly', () => {
    render(<BasicLogin onSubmit={() => {}} />);
    expect(screen.getByText('Welcome Back')).toBeInTheDocument();
    expect(screen.getByLabelText('Email')).toBeInTheDocument();
    expect(screen.getByLabelText('Password')).toBeInTheDocument();
    expect(screen.getByRole('button', { name: 'Submit' })).toBeInTheDocument();
  });

  it('submits correct data', () => {
    const mockSubmit = jest.fn();
    render(<BasicLogin onSubmit={mockSubmit} />);

    fireEvent.change(screen.getByLabelText('Email'), {
      target: { value: 'test@example.com' },
    });
    fireEvent.change(screen.getByLabelText('Password'), {
      target: { value: 'password123' },
    });
    fireEvent.click(screen.getByRole('button', { name: 'Submit' }));

    expect(mockSubmit).toHaveBeenCalledWith({
      email: 'test@example.com',
      password: 'password123',
    });
  });
});
